dos: add GetDeviceProcRelative
