+++ rom/dos/dos.conf
@@ -258,4 +258,5 @@ BOOL AssignAddToList(CONST_STRPTR name, BPTR lock, ULONG position) (D1, D2, D3)
 # AROS-specific
 BPTR LockRelative(BPTR lock, CONST_STRPTR name, LONG accessMode) (D1,D2,D3)
 BPTR OpenRelative(BPTR lock, CONST_STRPTR name, LONG accessMode) (D1,D2,D3)
+struct DevProc *GetDeviceProcRelative(BPTR lock, CONST_STRPTR name, struct DevProc *dp) (D1,D2,D3)
 ##end functionlist
