+++ rom/dos/getdeviceprocrelative.c
@@ -1,0 +1,376 @@
+/*
+    Copyright (C) 1995-2021, The AROS Development Team. All rights reserved.
+
+    Desc: GetDeviceProcRelative() - Find the filesystem for a path.
+*/
+
+#include <aros/debug.h>
+
+#include <proto/exec.h>
+#include <proto/utility.h>
+
+#include "dos_intern.h"
+
+extern struct Process *r(struct DeviceNode *dn, struct DosLibrary *DOSBase);
+static struct DevProc *deviceproc_internal(struct DosLibrary *DOSBase, BPTR baselock, CONST_STRPTR name, struct DevProc *dp);
+
+/*****************************************************************************
+
+    NAME */
+#include <proto/dos.h>
+
+        AROS_LH3(struct DevProc *, GetDeviceProcRelative,
+
+/*  SYNOPSIS */
+        AROS_LHA(BPTR, lock, D1),
+        AROS_LHA(CONST_STRPTR, name, D2),
+        AROS_LHA(struct DevProc *, dp, D3),
+
+/*  LOCATION */
+        struct DosLibrary *, DOSBase, 229, Dos)
+
+/*  FUNCTION
+        GetDeviceProcRelative() will search for the filesystem handler which
+        you should send a command to for a specific path.
+
+        By calling GetDeviceProcRelative() multiple times, the caller will be
+        able to handle multi-assign paths.
+
+        The first call to GetDeviceProcRelative() should have the |dp|
+        parameter as NULL.
+
+    INPUTS
+        lock - existing lock to compute filename from
+        name - Name of the object to find.
+        dp   - Previous result of GetDeviceProcRelative() or NULL.
+
+    RESULT
+        A pointer to a DevProc structure containing the information
+        required to send a command to a filesystem.
+
+    NOTES
+        This call is AROS-specific.
+
+    SEE ALSO
+        FreeDeviceProc()
+
+*****************************************************************************/
+{
+    AROS_LIBFUNC_INIT
+
+    struct DevProc *dp2;
+    
+    D(bug("[GetDeviceProcRelative] lock=0x%p name='%s' dp=0x%p\n", lock, name, dp));
+    dp2 = deviceproc_internal(DOSBase, lock, name, dp);
+
+#if DEBUG
+    bug("[GetDeviceProcRelative] = 0x%p", dp2);
+    if (dp2)
+        bug(", port=0x%p lock=0x%p dv=0x%p\n", dp2->dvp_Port, dp2->dvp_Lock, dp2->dvp_DevNode);
+    RawPutChar('\n');
+#endif
+
+    return dp2;
+
+    AROS_LIBFUNC_EXIT
+
+} /* GetDeviceProcRelative */
+
+static struct DevProc *deviceproc_internal(struct DosLibrary *DOSBase, BPTR baselock, CONST_STRPTR name, struct DevProc *dp)
+{
+    struct Process *pr = (struct Process *)FindTask(NULL);
+    struct DosList *dl = NULL;
+    char vol[32];
+    LONG len;
+    BPTR lock = BNULL;
+    BOOL res;
+    CONST_STRPTR origname = name;
+    struct FileLock *fl;
+
+    ASSERT_VALID_PROCESS(pr);
+
+    /* if they passed us the result of a previous call, then they want to
+     * loop over the targets of a multidirectory assign */
+    if (dp != NULL) {
+
+        /* if what they passed us is not a multidirectory assign, then there's
+         * nothing for us to do */
+        if (dp->dvp_DevNode != NULL &&
+            (dp->dvp_DevNode->dol_Type != DLT_DIRECTORY || !(dp->dvp_Flags & DVPF_ASSIGN))) {
+
+            /* cleanup */
+            if (dp->dvp_Flags & DVPF_UNLOCK)
+                UnLock(dp->dvp_Lock);
+
+            FreeMem(dp, sizeof(struct DevProc));
+            SetIoErr(ERROR_NO_MORE_ENTRIES);
+            return NULL;
+        }
+
+        /* it's fine, we'll start from here */
+        dl = dp->dvp_DevNode;
+
+        /* lock the dos list here, to match the result of the next block */
+        LockDosList(LDF_ALL | LDF_READ);
+    }
+
+    /* otherwise we need to find a place to start in the doslist based on the
+     * name they passed in */
+    else {
+        if (name == NULL || Strnicmp(name, "NIL:", 4) == 0) {
+            SetIoErr(ERROR_DEVICE_NOT_MOUNTED);
+            return NULL;
+        }
+
+         /* allocate structure for return */
+        if ((dp = AllocMem(sizeof(struct DevProc), MEMF_ANY | MEMF_CLEAR)) == NULL) {
+            SetIoErr(ERROR_NO_FREE_STORE);
+            return NULL;
+        }
+
+        /* something real, work out what it's relative to */
+        if (Strnicmp(name, "PROGDIR:", 8) == 0) {
+            lock = pr->pr_HomeDir;
+            /* I am not sure if these are correct but AOS does return
+             * non-NULL PROGDIR: handle even if pr_HomeDir is cleared */
+            if (!lock)
+                lock = pr->pr_CurrentDir;
+            if (lock) {
+                fl = BADDR(lock);
+                dp->dvp_Port = fl->fl_Task;
+                dp->dvp_Lock = lock;
+                dp->dvp_DevNode = BADDR(fl->fl_Volume);
+            } else {
+                dp->dvp_Port = DOSBase->dl_Root->rn_BootProc;
+                dp->dvp_Lock = BNULL;
+                dp->dvp_DevNode = NULL;
+            }
+            dp->dvp_Flags = 0;
+            return dp;
+        }
+
+        /* extract the volume name */
+        len = SplitName(name, ':', vol, 0, sizeof(vol) - 1);
+
+        /* if there wasn't one (or we found a lone ':') -> current dir */
+        if (len <= 1) {
+            lock = baselock;
+            if (!lock)
+                lock = pr->pr_CurrentDir;
+            /* if we got NULL, then it's relative to the system root lock */
+            if (lock != BNULL) {
+                fl = BADDR(lock);
+                dp->dvp_Port = fl->fl_Task;
+                dp->dvp_Lock = lock;
+                dp->dvp_DevNode = BADDR(fl->fl_Volume);
+            } else {
+                dp->dvp_Port = DOSBase->dl_Root->rn_BootProc;
+                dp->dvp_Lock = BNULL;
+                dp->dvp_DevNode = NULL;
+            }
+
+            /* Jump to the root directory if name[0] == ':' */
+            if (name[0] == ':')
+                dp->dvp_Lock = (dp->dvp_DevNode) ? dp->dvp_DevNode->dol_Lock : BNULL;
+
+            dp->dvp_Flags = 0;
+            return dp;
+        }
+ 
+        do {
+            /* now find the doslist entry for the named volume */
+            dl = LockDosList(LDF_ALL | LDF_READ);
+            dl = FindDosEntry(dl, vol, LDF_ALL);
+
+            /* not found, bail out */
+            if (dl == NULL) {
+                UnLockDosList(LDF_ALL | LDF_READ);
+
+                if (ErrorReport(ERROR_DEVICE_NOT_MOUNTED, REPORT_INSERT, (IPTR)vol, NULL) == DOSTRUE) {
+                    FreeMem(dp, sizeof(struct DevProc));
+                    return NULL;
+                }
+            }
+        } while(dl == NULL);
+    }
+
+    /* at this point, we have an allocated devproc in dp, the doslist is
+     * locked for read, and we have the entry for the named "volume"
+     * (device, assign, etc) in dl and a filename relative to that in name */
+
+    /* late assign. we resolve the target and then promote the doslist entry
+     * to full assign */
+    if (dl->dol_Type == DLT_LATE) {
+        /* obtain a lock on the target */
+        lock = Lock(dl->dol_misc.dol_assign.dol_AssignName, SHARED_LOCK);
+
+        /* didn't find the target */
+        if (lock == BNULL) {
+            UnLockDosList(LDF_ALL | LDF_READ);
+            FreeMem(dp, sizeof(struct DevProc));
+            return NULL;
+        }
+
+        /* Directly change assign type without calling AssignXXX(),
+         * mimics AOS behavior, AOS programs assume it is safe to
+         * keep LDF_WRITE lock while calling Lock("late assign:");
+         */
+        dl->dol_Type = DLT_DIRECTORY;
+        dl->dol_Lock = lock;
+        dl->dol_Task = ((struct FileLock*)BADDR(lock))->fl_Task;
+        FreeVec(dl->dol_misc.dol_assign.dol_AssignName);
+        dl->dol_misc.dol_assign.dol_AssignName = NULL;
+
+        /* the added entry will be a DLT_DIRECTORY, so we can just copy the
+         * details in and get out of here */
+        dp->dvp_Port = dl->dol_Task;
+        dp->dvp_Lock = dl->dol_Lock;
+        dp->dvp_Flags = 0;
+        dp->dvp_DevNode = dl;
+
+        UnLockDosList(LDF_ALL | LDF_READ);
+
+        return dp;
+    }
+
+    /* nonbinding assign. like a late assign, but with no doslist promotion */
+    if (dl->dol_Type == DLT_NONBINDING) {
+        lock = Lock(dl->dol_misc.dol_assign.dol_AssignName, SHARED_LOCK);
+
+        /* just fill out the dp and return */
+        dp->dvp_Port = ((struct FileLock *) BADDR(lock))->fl_Task;
+        dp->dvp_Lock = lock;
+        dp->dvp_Flags = DVPF_UNLOCK;   /* remember to unlock in FreeDeviceNode() */
+        dp->dvp_DevNode = dl;
+
+        UnLockDosList(LDF_ALL | LDF_READ);
+
+        return dp;
+    }
+
+    /* devices and volumes are easy */
+    if (dl->dol_Type == DLT_DEVICE || dl->dol_Type == DLT_VOLUME)
+    {
+        struct MsgPort *newhandler = NULL;
+
+        if (dl->dol_Type == DLT_DEVICE)
+        {
+            /* Check if the handler is not started */
+            newhandler = dl->dol_Task;
+            if (!newhandler)
+            {
+                D(bug("[GetDeviceProcRelative] Accessing device '%b', path='%s'\n", dl->dol_Name, origname));
+
+                /*
+                 * Unlock before starting handler, handler may internally
+                 * require dos list locks, for example to add volume node.
+                 */
+                UnLockDosList(LDF_ALL | LDF_READ);
+
+                newhandler = RunHandler((struct DeviceNode *)dl, origname, DOSBase);
+                if (!newhandler)
+                {
+                    FreeMem(dp, sizeof(struct DevProc));
+                    SetIoErr(ERROR_DEVICE_NOT_MOUNTED);
+                    return NULL;
+                }
+
+                LockDosList(LDF_ALL | LDF_READ);
+            }
+        }
+        else
+        {
+            res = TRUE;
+
+            while (res && !dl->dol_Task)
+            {
+                D(bug("[GetDeviceProcRelative] Accessing offline volume '%b'\n", dl->dol_Name));
+                res = !ErrorReport(ERROR_DEVICE_NOT_MOUNTED, REPORT_VOLUME, (IPTR)dl, NULL);
+            }
+
+            if (!res)
+            {
+                UnLockDosList(LDF_ALL | LDF_READ);
+                FreeMem(dp, sizeof(struct DevProc));
+                SetIoErr(ERROR_DEVICE_NOT_MOUNTED);
+                return NULL;
+            }
+        }
+
+        /*
+         * A handler theoretically may choose to use custom MsgPort for communications.
+         * Pick up its preference if specified.
+         */
+        dp->dvp_Port = dl->dol_Task ? dl->dol_Task : newhandler;
+        dp->dvp_Lock = BNULL;
+        dp->dvp_Flags = 0;
+        dp->dvp_DevNode = dl;
+
+        UnLockDosList(LDF_ALL | LDF_READ);
+
+        return dp;
+    }
+
+    /* sanity check */
+    if (dl->dol_Type != DLT_DIRECTORY) {
+        UnLockDosList(LDF_ALL | LDF_READ);
+        FreeMem(dp, sizeof(struct DevProc));
+        kprintf("%s:%d: DosList entry 0x%p has unknown type %d. Probably a bug, report it!\n"
+                "    GetDeviceProcRelative() called for '%s'\n",
+                __SRCFILENAME__, __LINE__, dl, dl->dol_Type, name);
+        SetIoErr(ERROR_BAD_NUMBER);
+        return NULL;
+    }
+
+    /* real assigns. first, see if it's just pointing to a single dir */
+    if (dp->dvp_Flags != DVPF_ASSIGN) {
+        /* just a plain assign, easy */
+        dp->dvp_Port = dl->dol_Task;
+        dp->dvp_Lock = dl->dol_Lock;
+        dp->dvp_DevNode = dl;
+        
+        /* note multidirectory assigns so the caller knows to loop */
+        dp->dvp_Flags = dl->dol_misc.dol_assign.dol_List != NULL ? DVPF_ASSIGN : 0;
+
+        UnLockDosList(LDF_ALL | LDF_READ);
+        return dp;
+    }
+
+    /* finally the tricky bit - multidirectory assigns */
+
+    /* if we're pointing at the "primary" lock, then we just take the first
+     * one in the list */
+    if (dp->dvp_Lock == dl->dol_Lock)
+        dp->dvp_Lock = dl->dol_misc.dol_assign.dol_List->al_Lock;
+
+    /* otherwise we're finding the next */
+    else {
+        struct AssignList *al = dl->dol_misc.dol_assign.dol_List;
+
+        /* find our current lock (ie the one we returned last time) */
+        for (; al != NULL && al->al_Lock != dp->dvp_Lock; al = al->al_Next);
+
+        /* if we didn't find it, or we didn't but there's none after it, then
+         * we've run out */
+        if (al == NULL || (al = al->al_Next) == NULL) {
+            UnLockDosList(LDF_ALL | LDF_READ);
+            FreeMem(dp, sizeof(struct DevProc));
+
+            SetIoErr(ERROR_NO_MORE_ENTRIES);
+            return NULL;
+        }
+
+        /* fill out the lock from the new entry */
+        dp->dvp_Lock = al->al_Lock;
+    }
+
+    /* final pieces */
+    dp->dvp_Port = ((struct FileLock *) BADDR(dp->dvp_Lock))->fl_Task;
+    dp->dvp_Flags = DVPF_ASSIGN;
+    dp->dvp_DevNode = dl;
+
+    UnLockDosList(LDF_READ|LDF_ALL);
+    /* phew */
+    SetIoErr(0);
+    return dp;
+}
